requirements.txt
psycopg2
pandas
numpy
python-dotenv
pymango
certifi
pymongo[srv]
dill
scikit-lean
pyaml
mlflow
#-e .

